+++ src/Core/Query/FieldSets.php
@@ -71,8 +71,8 @@ class FieldSets implements Arrayable, IteratorAggregate, Countable
             throw new \InvalidArgumentException('Expecting an array or enumerable object.');
         }
 
-        return new self(...collect($value)->map(function (array $fields, string $type) {
-            return new FieldSet($type, ...$fields);
+        return new self(...collect($value)->map(function ($fields, string $resourceType) {
+            return FieldSet::cast($resourceType, $fields);
         })->values());
     }
 
@@ -127,6 +127,17 @@ class FieldSets implements Arrayable, IteratorAggregate, Countable
     }
 
     /**
+     * Get a field set by resource type.
+     *
+     * @param string $resourceType
+     * @return FieldSet|null
+     */
+    public function get(string $resourceType): ?FieldSet
+    {
+        return $this->stack[$resourceType] ?? null;
+    }
+
+    /**
      * @return bool
      */
     public function isEmpty(): bool
@@ -147,9 +158,9 @@ class FieldSets implements Arrayable, IteratorAggregate, Countable
      */
     public function toArray()
     {
-        return collect($this->stack)->map(function (FieldSet $fieldSet) {
-            return $fieldSet->fields();
-        })->all();
+        return collect($this->stack)
+            ->map(fn(FieldSet $fieldSet) => $fieldSet->toString())
+            ->all();
     }
 
     /**